+++ lib/presentations/features/home/home_view.dart
@@ -1,6 +1,7 @@
 import 'package:easy_localization/easy_localization.dart' as local;
 
 import 'package:animated_theme_switcher/animated_theme_switcher.dart';
+import 'package:ebla/presentations/features/rent/widgets/rent_grid_item_widget.dart';
 import 'package:ebla/presentations/resources/assets_manager.dart';
 import 'package:ebla/presentations/resources/theme_manager.dart';
 import 'package:ebla/presentations/widgets/growth_rate_widget.dart';
@@ -33,29 +34,7 @@ class HomeView extends StatefulWidget {
 
 class _HomeViewState extends State<HomeView> {
   // bool _switchValue = false;
-  //todo: this should be removed once we have real data from api
-  List<GridItemData> gridItemsData = const [
-    GridItemData(
-        title: "إجمالي عدد  \nعقود الإيجار",
-        value: "983",
-        imagePath: 'imagePath',
-        valueUnit: ''),
-    GridItemData(
-        title: "إجمالي عدد الوحدات \n العقارات المستأجرة \\",
-        value: "983",
-        imagePath: 'imagePath',
-        valueUnit: ''),
-    GridItemData(
-        title: "متوسط قيمة الإيجار\n لكل وحدة \\ عقار",
-        value: "5,790",
-        imagePath: "imagePath",
-        valueUnit: 'ر.ق'),
-    GridItemData(
-        title: "إجمالي قيمة عقود\n الإيجار",
-        value: "716,104",
-        imagePath: "imagePath",
-        valueUnit: 'ر.ق'),
-  ];
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -112,83 +91,7 @@ class _HomeViewState extends State<HomeView> {
                       mainAxisSpacing: AppSizeH.s22,
                       crossAxisSpacing: AppSizeW.s23,
                       gridItemChildBuilder: (context, index) {
-                        return Container(
-                          decoration: BoxDecoration(
-                            borderRadius:
-                                BorderRadius.all(Radius.circular(AppSizeR.s20)),
-                            boxShadow: [
-                              BoxShadow(
-                                  offset: const Offset(1, 1),
-                                  spreadRadius: AppSizeR.s2,
-                                  blurRadius: AppSizeR.s11,
-                                  color: ColorManager.black.withAlpha(30)),
-                            ],
-                            gradient: LinearGradient(
-                              colors: [
-                                ColorManager.platinum,
-                                ColorManager.white,
-                              ],
-                              begin: Alignment.topLeft,
-                              end: Alignment.bottomRight,
-                              stops: const [0.2, 1.0],
-                            ),
-                          ),
-                          child: Column(children: [
-                            SizedBox(height: AppSizeH.s16),
-                            Row(
-                              children: [
-                                Expanded(
-                                  child: Text(gridItemsData[index].title,
-                                      textAlign: TextAlign.center,
-                                      style: TextStyle(
-                                          fontSize: AppSizeSp.s13,
-                                          fontWeight: FontWeight.w700)),
-                                ),
-                              ],
-                            ),
-                            SizedBox(
-                              height: AppSizeH.s14,
-                            ),
-                            Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                SizedBox(width: AppSizeW.s28),
-                                Flexible(
-                                  child: Wrap(
-                                    alignment: WrapAlignment.center,
-                                    children: [
-                                      FittedBox(
-                                        child: Text(
-                                          gridItemsData[index].value,
-                                          style: TextStyle(
-                                              fontSize: AppSizeSp.s18,
-                                              fontWeight: FontWeight.w700),
-                                        ),
-                                      ),
-                                      Text(
-                                        gridItemsData[index].valueUnit,
-                                        style: TextStyle(
-                                            fontSize: AppSizeSp.s18,
-                                            fontWeight: FontWeight.w700),
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                                Flexible(
-                                  child: Image.asset(
-                                    matchTextDirection:
-                                        Directionality.of(context) ==
-                                            TextDirection.ltr,
-                                    ImageAssets.soldOrRentedUnits,
-                                    height: AppSizeH.s65,
-                                    width: AppSizeW.s65,
-                                  ),
-                                ),
-                              ],
-                            )
-                          ]),
-                        );
+                        return RentGridItemWidget(index: index);
                       },
                     ),
                   ),
@@ -270,17 +173,4 @@ class _HomeViewState extends State<HomeView> {
       ),
     );
   }
-}
-
-class GridItemData {
-  final String title;
-  final String value;
-  final String imagePath;
-  final String valueUnit;
-
-  const GridItemData(
-      {required this.title,
-      required this.value,
-      required this.imagePath,
-      required this.valueUnit});
 }
